refactor classifier validators
